// src/components/Dialog/Dialog.Types.ts
export interface DialogStoreProps {
    id?: string;
    isOpen?: boolean;
    isFromDialog?: boolean;

    variant?: "information" | 'warning' | 'error' | 'destructive' | 'confirmation';

    title?: string;
    icon?: 'success' | 'error' | 'warning' | 'info';
    description?: string;
    custom?: React.ReactNode;

    cancelAction?: () => void | Promise<void>;
    cancelActionLabel?: string;

    PrimaryAction?: () => void | Promise<void>;
    PrimaryActionLabel?: string;
    isPrimaryActionDisabled?: boolean;

    SecondaryAction?: () => void | Promise<void>;
    SecondaryActionLabel?: string;
    isSecondaryActionDisabled?: boolean;

    isDismissable?: boolean;
    isKeyboardDismissDisabled?: boolean;
}


export interface DialogLibProps {

}